cparser: nolib uses local rt; fixes

+++ cparser/ast_t.h
@@ -263,7 +263,8 @@ struct compound_literal_expression_t {
 	type_t            *type;
 	initializer_t     *initializer;
 	bool               global_scope;
-    unsigned short     id;
+    /* ast2firm data */
+    int                id; // sequence number for global, offset for local
 };
 
 struct builtin_constant_expression_t {
